Improve logging in src/layman/layer/geoserver/util.py::w[mf]s_direct
